skeleton

+++ src/main/java/com/intraedge/cc/daos/UserDao.java
@@ -19,7 +19,7 @@ public interface UserDao {
 	 * @param user the user entity
 	 * @return row id of the entity in the database
 	 */
-	public int createUser(User user) throws DataAccessException;
+	public long createUser(User user) throws DataAccessException;
 
 	/**
 	 * Change the password for a user
